apps/llm_agent/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
llama-cpp-python==0.2.27
python-dotenv==1.0.0
aiohttp==3.9.1
python-multipart==0.0.6